agent/requirements.txt
livekit-agents>=0.11
asyncio
python-dotenv
livekit
pillow
openai